Move new files to the right location
